Filter dashboard cards by student
A new parameter has been added to the dashboard_cards API to filter
cards by the associated user, if the parameter sent is the same user
in session, the API will return only course related to the user,
excluding the observer enrollments. All the request responses will be
saved in the K5Dashboard component state, this will prevent the
dashboard to make requests if they have been done previously

closes LS-2478
flag = k5_parent_support

Test Plan:

Step 1: Create 3 courses and 3 users and do the following

In course 1
Assign student role to the user 1 and user 3
Assign observer role to user 2 and link the user to the user 1

In course 2
Assign student role to user 1.
Assign observer role to user 3 and link the user to user 1

In course 3
Assign teacher role to the user 3

Step 2: with the parent support flag enabled, log in as each
user and expect to see the following when viewing the dashboard/home

For user 1:
- Cards for course 1 and 2 to be shown

For user 2:
-Only the card for course 1 to be shown

For user 3:
- Cards for course 1 and 3 to be shown, as the user picker selects
himself by default
- Go to the user picker and select the user 1 and expect
- To see the card for course 2 be rendered
- Open the browser console and go to the network and notice that no
requests to the dashboard_cards API are made when switching user in
the picker, as both options responses are stored in the component
state now

Step 3: Disable the parent support mode and expect to see all the
enrollments for each user as usual

Change-Id: I1bfd0bd38e00be09f0e72ebb22a68825d1e85464
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/271780
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Peyton Craighill <pcraighill@instructure.com>
Reviewed-by: Eric Saupe <eric.saupe@instructure.com>
QA-Review: Eric Saupe <eric.saupe@instructure.com>