Tweak how canvas-planner uses responsiviser
This commit settles on 1 responsiviser at the very
top of the component tree and passes responsiveSize down
as a prop. Before, when each of the components interested
in the responsiveSize were wrapped in the responsiviser,
you'd get 2 renders with 2 different sizes as the change
propagaged. This way when the breakpoint is hit, there's
one render update to the whole tree.

My original idea was to use react context to pass the
responsiveSize down the tree, but the redux Provider
uses the old context api and react will not premit
both to be present.

Also update babel config to quiet all the warnings.

closes: LS-2550
flag=none

test plan:
  - the planner adjusts its layout as the screen width
    shrinks just like it did before

Change-Id: I8d6e90f5dfee5ec47f9abce18d7626f09f96f293
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/271752
Reviewed-by: Jeff Largent <jeff.largent@instructure.com>
QA-Review: Jeff Largent <jeff.largent@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>