Keep course dates in sync with term enrollment setting
When `restrict_enrollments_to_course_dates` is false for a course, and the
course is associated with a term, enrollment is to be restricted to the
dates of the term. This change ensures that `start_at` and `conclude_at`
are not set (and are cleared if necessary) in this state.

The only exception is the `start_at` for a course which is already
published. This value is allowed to exist, as publishing a course writes
an immediate start date.

fixes LS-2553
flag=none

test plan:
- create a new course which is associated with a term and has "Term"
  participation selected on the course settings page
- attempt to use the `update` API endpoint to set a `start_at` or
  `conclude_at` date for the course
- verify that neither of these properties is updated
- change the participation setting to "Course" on the course settings page
- re-send your API request, and verify that the dates were updated as
  desired
- attempt to use the `create` API endpoint to create a new course without
  specifying a value for `restrict_enrollments_to_course_dates`, causing
  it to default to false. be sure to include `start_at` and `conclude_at`
  values in your request
- verify that `start_at` and `conclude_at` are nil on the created course
- repeat the previous `create` API step, but set
  `restrict_enrollments_to_course_dates` to true
- verify that `start_at` and `conclude_at` are not nil on the created
  course

Change-Id: I230c3d17c9db638b7c6e09997a168660ac5abcb2
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/271928
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeremy Stanley <jeremy@instructure.com>
QA-Review: Jeremy Stanley <jeremy@instructure.com>
Product-Review: Isaac Moore <isaac.moore@instructure.com>