Observer support for Grades tab
We thought filtering the dashboard cards will automatically update
the grades tab, but the info used in the tab is fetched when the
GradesPage is mounted and it does not depend on the dashboard card
at all, so to fix this the observedUserId is now synced with the tab
anytime the observee picker changes, the courses in the GradesTab
component state will be filtered by the associated user, and the
grades information will be calculated based on the observee
enrollment to make sure the observer can see the grades too, if the
observed user is the same as the one in session, the observer
enrollments will be filtered out and only the enrollments of
students, teachers, etc. will be shown, this to keep consistency with
the dashboard cards.

Closes LS-2582
flag = k5_parent_support

Test plan:
- With the parent support feature enabled
- Log in as observer and expect to see the user picker at the top of
the dashboard
- Go to the Grades tab and expect to see only the courses for the
filtered student, expect to see the observee grades too
- Change the observed student in the picker and expect the Grades tab
be updated.

Change-Id: I0aa157d369b6be9abc131d12cdf3dbb8f7d5a36b
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/273259
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Peyton Craighill <pcraighill@instructure.com>
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>