Address date/time parsing issues in non-en locales
Refs FOO-2347
flag=none
[fsc-max-nodes=18] [fsc-timeout=30]

The old jQuery-based date picker was relying on the ability
to parse what it formats, which is far, far from a safe bet in
non-English locales. One example is that in French, a date in
September will format as, say, "1 Sep 2021" but that does not
parse properly (it has to be "1 Sept 2021" which is the French
abbreviation for the month). This caused no end of trouble in
those date pickers.

This modifies the DatetimeField and the jQuery date picker
so that it always uses a stored ISO string date (which is
guaranteed to be able to be parsed by new Date()), thus
eliminating the parsing errors when the pickers are used.
The only exception is when the field is being typed into,
which still uses tz.parse and also includes a running value
underneath the field to help the user get the format right.

Test plan:
* Go to an assignment editor (/courses/:id/assignments/:id/edit)
* Scroll down to the dates at the bottom
* Those date and time pickers should all still work normally
* Try it in the fr locale, especially selecting a date in the
  month of September and setting the time to 11:11
* It should still work!
* Try other locales, too.

Change-Id: Idd4ba1684dc4e3679ec2360ffa4f362924756103
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/273803
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ahmad Amireh <ahmad@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
QA-Review: Ahmad Amireh <ahmad@instructure.com>