skipping needed but broken test
The test in CanvasInboxFullPage.test.js was skipped
because it was passing based on a false poitive and
now fails with changes made in g/273709. The way
that the msw mocks are written overwrites the change
triggered by the mark as read call, nullifying the
behavior we were trying to test.

VICE-2077 is the plan to fix and unskip this test
some time in the future.

refs VICE-1128
flag=react_inbox

Test Plan:
-specs should pass

Change-Id: I80b3fd3996a85f7a510bc756a51dfcdbbac7ca75
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/273804
Reviewed-by: Jeffrey Johnson <jeffrey.johnson@instructure.com>
QA-Review: Jeffrey Johnson <jeffrey.johnson@instructure.com>
Product-Review: Jeffrey Johnson <jeffrey.johnson@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>