fix draft caching causing re-renders
this update cache method was a copy from the update cache on new entries
the entries needs to put the entry in a specific spot on the page, but
drafts can just go into the cache

in this case we had zero entries and then it was doing -1, so the page
ended trying to load page -1, which would just fail and hit the catch.
the catch would cause a re-render since it refetches the discussion
object.

Test Plan:
 - have a discussion topic with zero replies
 - go to reply to the message
 - when the draft saves it should update the cache and not re-render
 - if there were an error it would not re-render the page
 
refs VICE-2079
flag=react_discussions_post

Change-Id: I1ab050ae0b9ef7cab6bf0a20dd0a26df41a73f7b
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/274451
Reviewed-by: Davis Hyer <dhyer@instructure.com>
QA-Review: Davis Hyer <dhyer@instructure.com>
Product-Review: Davis Hyer <dhyer@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>