let apollo graphql handle where to put new nodes for a connection
the page param was being set to specify where in the cache a node should
go, but all we need is the first or last spot in the nodes on the query.

the 'page' variable param is also a poorly named variable that is a page
reference or marker. The query uses the page variable to know to get the
entries after a specific position within the discussion entry connection

this was causing extra queries almost every time because of the refetch.

test plan
 - reply to a discussion it should show up on the first page at the top.
 - it should not run an extra query and should cause error in the catch.
 - sort entries newest last and reply to discussion it should change you
   to the last page, highlight the entry at the bottom of the last page.
 - it should do 1 query to get the last page, and no additional queries.

fixes VICE-2094
fixes VICE-2106
flag=react_discussions_post

Change-Id: I63d9a2869055b9d8c48ca85445f3f91305728d9e
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/275189
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Drake Harper <drake.harper@instructure.com>
QA-Review: Drake Harper <drake.harper@instructure.com>
Product-Review: Drake Harper <drake.harper@instructure.com>