Correct for ENV.TIMEZONE vs browser's timezone
Refs FOO-2347
flag=none

A previous commit used Date methods to construct an ISO8601
string to pass back to DatetimeField when a selection is
made in the date picker. Unfortunately it didn't take into
account that all operations on a Date are done in the timezone
of the BROWSER, which is not necessarily what ENV.TIMEZONE
is set to. We have to correct between the two.

This bug was not found in QA before because most people have
their Canvas timezone set to match their local timezone.

Test plan:
* Set your Canvas user time zone to one west of where you are
  i.e. if you're in US Central, set it to US Mountain
* Edit an assignment (/courses/:id/assignments/:id/edit)
* Open the date picker for the due date and click around on
  various dates
* The time should remain the same unless you change it too
* Remember what you've changed it to and then click Save
* The date and time you just selected should be reflected in
  the "Due" column in the assignment screen

Change-Id: I5c4b1053935dab48ff5b202e02adf5a94ad1deae
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/275285
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Ahmad Amireh <ahmad@instructure.com>
QA-Review: Ahmad Amireh <ahmad@instructure.com>
Product-Review: Ahmad Amireh <ahmad@instructure.com>