Manually set the value of dev key toggle
Rather than having the radio button group keep track of its
own value, we need to manually manage that value. This will allow
us to show a confirmation modal before changing the value, so that
we can sometimes *not* change the value if the user clicks "no" on
the modal.

Because the state object now updates right when the toggle is clicked,
(previously it only updated when the API call returned),
we need to set the state back if the API call fails. This has the nice
side effect of setting the toggle switch back to its original value
if the API call fails. That is why setBindingWorkflowStateSuccessful and
setBindingWorkflowStateFailed now actually do something instead of just
being a no-op; the dev key list needs to actually update with new value
from the API call, instead of just flipping the toggle switch and
sending the API request in the background.

The radioGroupValue function now returns the actual *displayed* value of
the RadioInput. Previously, for inherited keys, it would return "allow"
and then have the getDefaultValue function interpret that as "off."

test plan:
- this is probably easier to test along with the commit after this one, which
  adds the confirmation modal
- developer keys can be turned on and off from the developer keys page
- inherited developer keys default to "off," can be enabled, and stay
  enabled after a page refresh

flag = none
refs INTEROP-7091

Change-Id: Ib74a104ce3be4657c57108ec7430320995a3feb8
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/275781
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Tucker Mcknight <tmcknight@instructure.com>