Only show Send|Copy To in course files
The code for handling Send To and Copy To are in the
courses_controller, so let's only show the menuitems
when in course Files and not user files (or group files,
if that's a thing).

Also converted the QUnit specs to jest.

Changed ItemCog string refs (which were only used in specs
except for 1 case) to data-testid

closes LS-2724
flag=none

test plan:
  - Send To... and Copy To... work as expected in course files
  - Neither menu item is available in user files
   (you find the menu in the kabob menu to the right of a file
   on the files page)

Change-Id: I89f3d3706645c183381523d07956e692069dc0cc
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/275792
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Jeff Largent <jeff.largent@instructure.com>
QA-Review: Jeff Largent <jeff.largent@instructure.com>
Product-Review: Ed Schiebel <eschiebel@instructure.com>