add account rights check when no associated accounts
In rare circumstances if a user doesn't belong to an account
and is added to a course when enrolling users, they will be in
a "creation_pending" workflow state. In this case the user doesn't
have any associated accounts, thus when a user (admin) is
attempting to seek permission on a user i.e. :read_full_profile
it will always be false.

fixes FOO-2514
flag = none

Test plan:
• Create a course and enroll students who have never before had
  an account so that they show as a pending enrollment
  • NOTE: you must add the user via Course -> People -> |+People|
    so they're added with a workflow_state of 'creation_pending'
• Masquerade or log in as an account admin
  (with all permissions enabled) and go to the People tab of the
  course and click on an unconfirmed student enrollment
• Notice the context card no longer flashes with a javascript error
  and the context card is displayed like all other users within
  the specified course

Change-Id: I33049bd9153031e78f36b70b9ea8005b210b44c2
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/277279
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Cody Cutrer <cody@instructure.com>
QA-Review: August Thornton <august@instructure.com>
Product-Review: August Thornton <august@instructure.com>