RuboCop: Lint/LiteralAsCondition
[skip-stages=Flakey]

all manual

the majority of these were negating symbols to document a boolean
param. the preferred way to do this is to use a kwarg instead.
some of them were converted to that, but some there were many
callsites and conflicting definitions so I didn't want to go that
far in this commit, and I just changed to `true` or `false`
as appropriate

Change-Id: Iccfbc1c7a1fb744704343130a6b2951f6741192f
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/277547
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Simon Williams <simon@instructure.com>
QA-Review: Cody Cutrer <cody@instructure.com>
Product-Review: Cody Cutrer <cody@instructure.com>