MSFT sync: delete mappings for non-existent users
Previously, we never deleted the mappings between a Canvas User and
Microsoft user AAD object ID after caching it. So if a Microsoft admin
deleted users from Microsoft 365, some courses would fail to sync
because the AAD object ID we had didn't exist on the Microsoft side, and
ge would never try to re-look it up.

Now, if we try to add a user and the user's AAD object ID doesn't exist
on the Microsoft side, we will notice that and delete the mapping. The
next time a sync is run, we will successfully re-lookup the Canvas user
and add them if a different Microsoft user exists for them.

If we are trying to remove a user from a group and the user doesn't
exist, we will still error, as Microsoft's API doesn't differentiate
between the user not existing, group not existing, or the user just not
being in the group. However this case is not as bad as it can only
happen in a partial sync (since in a full sync we get the full list of
users in the group from the API, and that list will not include deleted
users). So if that happened, a full sync would fix the course/MS group
and allow it to sync again.

This commit:
* Modifies add_users_to_group (patch api -- all users at once) to notice
  when the API sends back a 404 that some user does not exist. At this
  point we only know that at least one user doesn't exist, so we use the
  batch API (all users as separate requests in a batch) to get a list of
  all such users.
* Changes add_users_to_group_via_batch to notice when the API returns a
  404 saying the user does not exist. To communicate this up, we have
  a new method nonexistent_users on GroupMembershipChangeResult
* Changes special_case.rb to support a block tester. Also use an
  encapsulate response which can contain a `batch_request_id` for batch
  requests. These are used by the new code.

closes INTEROP-7000

Test plan:
* have a course with some users (enrollments must be 'active',
  CommunicationChannels must be 'active' if email is used for lookup
  key) and sync with Microsoft.
* delete users on MS side
* run a full sync (e.g. group.syncer_job.run_synchronously). sync should
  succeed, the usermapping for the user that no longer has a user on the
  Microsoft side should be deleted.
* delete the group on MS side.
* add a user (active enrollment, active communicationchannel as before)
  with a user on the microsoft side and run a full sync. that should
  should recreate the group on the microsoft side (as before)
* delete a user on the MS side and then immediately remove the user from
  the course on the Canvas side. run a partial sync (e.g.
  group.syncer_job.run_synchronously(:partial)). That should be ignored
  (as before) because we just assume the user is not in the group (we
  can't distinguish between that and the user not existing when removing
  users).

Change-Id: I5230b32d7cbfd1462177e3e597ba7d9633f055d6
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/278068
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Evan Battaglia <ebattaglia@instructure.com>