Fix regression in jQuery datepicker
Fixes FOO-2564
Refs LS-2740
flag=none

g/276168 introduced several regressions into the functioning of
the jQuery datepicker. In particular, it re-introduced the use
of tz.parse to pull the current value back out of the field
string value, which is extremely error-prone in non-en-US locales.
It also broke the "running value" display while a user types in a
free-format value into the field.

This basically reverts most of that commit and implements the fix
in a different way.

Test plan:
* Review "steps to reproduce" in the FOO-2564 ticket
* The picker should now work right
* Re-run the test plan for g/276168 to make sure that this fix
  didn't re-introduce that bug (I did check this myself)

Change-Id: I7b9c177a84269ba8c410eafd751945f8055d3846
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/278202
Reviewed-by: Ed Schiebel <eschiebel@instructure.com>
QA-Review: Ed Schiebel <eschiebel@instructure.com>
Product-Review: Charley Kline <ckline@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>