Add message when student doesn't have first or last name
This message only is used in gradebook to show that the student
name does not consist of a first or last name. This is not intended
for gradebook exports/imports.

closes EVAL-2040
flag=gradebook_show_first_last_names

Test plan:
With a teacher and a gradebook with some students (one of them has no
last name):
1. Enable the FF and the account setting for this feature
2. Navigate to gradebook and select "Split Student Names" view option
3. See that the gridbook shows separate last and first name columns
4. See that the student with no last name shows "<No last name>"
5. All other students show correct first and last names
6. Export gradebook and make sure the student with no last name
does not have the "<No last name>" string in the csv (should be
empty string)

Change-Id: I80f98739188878915ba9e9fb957f56d54eae65d2
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/280522
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Spencer Olson <solson@instructure.com>
Product-Review: Jody Sailor
Reviewed-by: Dustin Cowles <dustin.cowles@instructure.com>
Reviewed-by: Kai Bjorkman <kbjorkman@instructure.com>