Fix duplicate tools showing up on Apps page
fixes INTEROP-6572

The problem is that certain places in the UI flow call store.reset().
Before we kick off fetching of the list of tools, we check the store's
state's isLoading and isLoaded to make sure we don't fetch multiple
times. However, store.reset() resets the isLoading and isLoaded. When
the JSON requests/responses are interwoven with resets, we can have
multiple responses concatenating the tools multipe times. Example:

  store reset, externalTools = []
  XHR request (lti_apps) 1 made
  store reset, externalTools = []
  XHR request (lti_apps) 2 made
  XHR response 1 received; externalTools = [a, b, c]
  XHR response 2 received; externalTools = [a, b, c, a, b, c]

The most common call to reset() is when showing the App Center grid of
tools (see ExternalAppsStore.reset() in AppCenterStore.js). reset() is
also called when adding or editing a tool, but it is hard to do that
multiple times quickly before the lti_apps endpoint can return. Also
note that the list of tools seems to be cached because the lti_apps
endpoint can return a 304 (no new content), so initial requests to fetch
the list of tools are usually quicker.

I believe the other usages of $.ajax() in ExternalAppsStore are OK to
leave as-is. They are only things relating to one item (tool), and the
changes we might want to try to attempt anyway. And, importantly, they
don't concatenate to the list of tools as a successful retrieval of the
tools does.

This also fixes a mostly unrelated-bug where we were calling
store.fetch() without a store.reset() upon LTI 1.1 tool addition
failure, which again was causing us to show all the tools duplicated.

Test Plan:
- First, make sure you can repro the problem from the Account or Course
  Settings page. There are two ways:
  - Easiest: Go to the "Apps" tab and place your cursor over "View App
    Configurations"/"View App Center". Click your mouse button rapidly
    so it goes between to two (App center, and Configurations list) as
    fast as possible. There should then be duplicates.
  - More in line with Mysti's "switch tabs" way: edit the #index method
    in app/controllers/lti/lti_apps_controller.rb and add a "sleep 3"
    (e.g.). On the Apps tab, make sure it shows the App Center logos
    grid, not the installed tools list. Switch back and forth between
    the Apps tabs and another tab as quick as possible. This will force
    multiple reloads of the App Center. Then click "View App
    Configurations" and wait a few seconds. Duplicates should start
    appearing.
- Make sure that you cannot repro the bug after checking out this commit
  and building assets.
- For the second bug, also make sure you can repro it before and not
  after this commit:
  - Try to add an LTI 1.1 tool with no shared secrets (e.g. using Manual
    Entry when adding the tool).
  - Both before and after this commit, it should fail. But before this
    commit, it will also duplicate the currently-shown list of
    already-installed apps.

Change-Id: Ia99a7e46b895807822a0b430e2878c7ffe4ac848
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/282586
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Product-Review: Evan Battaglia <ebattaglia@instructure.com>