Include concluded section overrides for concluded enrollments
When a student had two section enrollments, one active and one concluded,
with the concluded section being assigned a quiz it was possible for the
student to still take the quiz even if the availabilty was in the past.
This was due to the quiz being overriden for the user but the concluded
enrollment not being included when checking if it is locked. This allowed
the user to see the quiz and not see that it is locked.

fixes LS-2895
flag=none

test plan:
- Create a course with 2 sections
- Create a quiz with access dates in the past assigned to only one section
- Create a student in the course with an active enrollment in one section
 and a concluded or inactive enrollment in the other section
 (this should be the section assigned to the quiz)
- Create a module and add the quiz to it
- As the student, navigate to the modules page and see that you are not
 able to take the quiz

Change-Id: I66b97ee60537599f852ad6dc3f2a270fa6e26eb2
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/282797
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Product-Review: Eric Saupe <eric.saupe@instructure.com>
Reviewed-by: Robin Kuss <rkuss@instructure.com>
QA-Review: Robin Kuss <rkuss@instructure.com>