Consider "Enter Grades as" setting with grading schemes
Do not match grading scheme values if the Gradebook setting for
"Enter Grades as" is set to "Points". See the community post in
the ticket to get a better understanding of the issue.

closes EVAL-2117
flag=none

Test Plan:
As a teacher:
1. Select/create a grading scheme that includes numbers e.g.
Name                         Range
5                            100% to 90.0%
4                            < 90.0% to 70.0%
3                            < 70.0% to 50.0%
Revision
required/Komplettering       < 50.0% to 25.0%
U                            < 25.0% to 0.0%
2. Create an assignment worth 100 points and set "Display grade as"
to Letter grade
3. In gradebook, grade the assignment for students using names in
grading scheme and other values  and see that totals correspond to
those values e.g. (5 -> 100%), (4 -> 89%) and (6 -> 6%)
4. Switch "Enter Grades as" option to "Points". Award 5 points to
a student and see that it does not assign the value from grading
scheme but the actual percentage. Try different values matching and
not matching grading scheme names.

Change-Id: I1312c6bfb5070c0b08e7590232e86d609e446025
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/283437
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Dustin Cowles <dustin.cowles@instructure.com>
Reviewed-by: Spencer Olson <solson@instructure.com>
QA-Review: Eduardo Escobar <eduardo.escobar@instructure.com>
Product-Review: Jody Sailor